Tokenizing Words and Sentences with NLTK
Tokenizing - Splitting sentences and words from the body of text.

Token - Each "entity" that is a part of whatever was split up based on rules. For examples, each word is a token when a sentence is "tokenized" into words. Each sentence can also be a token, if you tokenized the sentences out of a paragraph.